src/test/java/com/explorevehicle/controller/VehicleControllerTest.java
package com.explorevehicle.controller;

import com.explorevehicle.service.VehicleService;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.mockito.ArgumentMatchers.anyMap;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@SpringBootTest
class VehicleControllerTest {

    private MockMvc mockMvc;

    @Mock
    VehicleService vehicleService;

    @InjectMocks
    VehicleController vehicleController;

    private String vehicleTypeName = "Trailer";

    @BeforeEach
    public void setup() {
        MockitoAnnotations.openMocks(this);
        mockMvc = MockMvcBuilders.standaloneSetup(vehicleController).build();
    }

    @Test
    void testEndPoint_success() throws Exception {

        when(vehicleService.groupingByVehicleByDataType(vehicleTypeName)).thenReturn(anyMap());

        MvcResult result = this.mockMvc.perform(
                get("/vehicle/" + vehicleTypeName)
                        .accept(MediaType.APPLICATION_JSON_VALUE)
                        .contentType(MediaType.APPLICATION_JSON_VALUE))
                .andDo(print())
                .andExpect(status().isOk())
                .andReturn();

        String actual = result.getResponse().getContentAsString();
        assertNotNull(actual);
    }

    @Test
    void getVehicleDetails_404() throws Exception {
        mockMvc.perform(get("/v1/vehicelDetails/")
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().is4xxClientError());
    }
}
